temporarily disable libxcrypt on musl-linux-any
Discussion at #947193. While no decision has been made there, the status
quo is that musl includes libcrypt.so and crypt.h. So in order to get
things build as far as possible, we can temporarily opt out of libxcrypt
here. This will break as soon as packages start to depend on
libcrypt1-dev.