refactor the musl-linux-any bootstrap
The musl bootstrap has fewer stages than the glibc bootstrap. In the
implementation, I opted to skip gcc stage2 and glibc stage2 for musl.

Now glibc was refactored to permit a reduced toolchain bootstrap as
well. As it happens, the stage to remove there is not stage2, but
instead stage1.

To make musl and glibc more similar, the musl branch will now use the
code for glibc stage2 instead of glibc stage1. The idea being that we
can at some point remove the whole stage1 for all libcs.