two-stage toolchain bootstrap
glibc "recently" gained the ability to bootstrap with one less stage.
The stage1 (libc headers and an empty libc.so) are no longer necessary.
A bare metal gcc can directly build the stage2 glibc. This commit
discards stage1.

musl-aspects: musl already needs just one stage and the code was
recently updated to use that path that is kept. Note that we don't have
to install linux-libc-dev explicitly as musl doesn't actually depend on
it. We'll need it later, but not here.

hurd-aspects: A little later than glibc, hurd was also converted to
permit the reduced bootstrap. libihash-dev was needed for libpthread,
but it no longer is. We can directly build glibc stage2 using hurd
stage1. Care must be taken while installing glibc's Build-Depends as
hurd-headers-dev:hurd-i386 presently produces a file conflict with
libc6-dev-i386:amd64.

With luck, this all works out nicely now. Possibly, some corner cases
are still buggy and need to be fixed up. hurd and glibc stages should
likely be renamed as we will no longer need hurd stage2 or glibc stage1.