drop gcc/multilib patch #862756 to be fixed in dpkg-cross
The bug is about wrong dependencies in -arch-cross packages not
generated by dpkg-cross. However, the relevant dependencies originate
from .shlibs files in dpkg-cross'ed libc packages. It was closed by
Matthias as invalid since a while. Indeed, the actual issue is within
dpkg-cross. There is a proper patch for dpkg-cros #955631 for that now
and we no longer need this.