improve our LIMITS_H_TEST
LIMITS_H_TEST continues to be wrong. The original value essentially
expands to:

    [ -f $(BUILD_SYSTEM_HEADER_DIR)/limits.h ]

For wdotap builds, BUILD_SYSTEM_HEADER_DIR expands to /sys-include. The
test fails and gcc stage3 misbuilds.

In the debian packaging a patch adds:

    -o -f $(BUILD_SYSTEM_HEADER_DIR)/$(MULTIARCH_DIRNAME)/limits.h

This ended up not helping as BUILD_SYSTEM_HEADER_DIR is not what we
expect it to be.

Now rebootstrap changed it to `:`. This value happens to break the
stage1 build as there limits.h actually does not exist. When gcc assumes
existence, cpp fails early and the cpp detection prefers using
`/lib/cpp` which causes lots of issues.

So we now change it to the very unportable
/usr/include/$(MULTIARCH_DIRNAME)/limits.h to have it work at all.
Unfortunately, this is not upstreamable.