fix binutils hppa64 build
TARGET=hppa64-linux-gnu does not work, because binutils attempts to
resolve it to a Debian architecture and there is no hppa64 Debian
architecture. As such, it only gets hppa (32).

To get around this, we use the native binutils build, which happens to
also produce a hppa64 package somewhat separately to the cross binutils.