attempt to fix ssp linking for musl
Alpine (and maybe others) add -lssp_nonshared when -fstack-protector is
added to provide the symbol __stack_chk_fail_local. At present, this is
an out-of-tree gcc patch carried by e.g. Alpine. It makes dash build as
it passes -fstack-protector to the link step.

A later version of this patch just defines LINK_SSP_SPEC as
"-lssp_nonshared", but doing so breaks gcc stage3 as it builds
libssp_nonshared.a too late.