work around the crypt.h move breaking gcc #1014375
gcc stage3's libsanitizer needs struct crypt_data from crypt.h, which is
built from libxcrypt, which needs a cross compiler for building, thus
forming a cycle. For now, we break this cycle by assuming that the size
of crypt_data is architecture-independent, which kinda is guarantueed by
libxcrypt. It's not really a final solution.