rewrite wdotap patch set on top of gcc-for-host changes
The merging of gcc-for-host caused a lot of changes to the gcc packaging
and made the wdotap patch set from cross-gcc-dev no longer work. It
required significant changes and quite some aspects could be dropped
resulting in a much smaller patch that is inlined in rebootstrap for
now. We'll see whether that works better or worse.