first attempt at fixing the c++ include search path
Compared to cross-gcc-dev, I dropped setting with_multiarch_cxxheaders
(part of patch 0005). As a result, g++-multiarch-incdir.diff is not
applied during build. Also the fixup for --with-gxx-include-dir (patch
0008) is not added to g++-multiarch-incdir.diff. Rather
--with-gxx-include-dir is being used.

Unfortunately, it has subtle semantics. In particular, the sysroot
(which happens to be "/") is stripped from the --with-gxx-include-dir
value. In effect, that turns our search path into a relative path. If
you run g++ from a directory other than /, it happens to not work
breaking the build of gmp.

By duplicating the leading / in --with-gxx-include-dir, we account for
the stripping of the sysroot and retain it as absolute (with or without
stripping). It probably isn't meant like this, but this seems to
practically work.

Many thanks to Samuel Thibault for providing much of the analysis
summarized herein.