work around binutils-for-host not being installable #1094189
binutils-for-host:target is built from the targeted cross build, but
that build does not emit binutils-common:target. Matthias changed
binutils-common such that there no longer are per-architecture
differences and it can be marked Multi-Arch: foreign, but we cannot mark
it that way here as we are not building native binutils. Hence work
around.