add a native pass for gcc-defaults
This is important when Debian's default gcc version differs from the gcc
version that we target as we now install the host's gcc-defaults patched
to target our desired gcc version. But since it has M-A:same components
that we actually need to coinstall, we also need the native gcc-defaults
to be patched in the same way.